Moving the part definition resolver to its matching namespace. Making sure that the part node resolver does not use its own DBContext.
